requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pymongo
python-dotenv